v0.5.8 changes for sql name escaping, select_irows() empty case.
Correct flattening operation so that strings like '[]' are converted to []. Same for dict and tuple.
Cleaned up a number of imports due to pyflakes linting
Fixed creation of index name in sql_utils and daf_sql due to change in name escaping.
Fixed edge case in select_irows() if self.lol is empty.