Changes to support AE in prep for dynamic HTML tables.
Improve initialization of disp_cols so it can be None, List or tuple and emitted as [] if not provided.
Added .to_donpa() which would convert specified columns to dict of numpy arrays of each col, where dict keys are col names.
This creates a simple pandas-like object where NumPy column operators can be easily applied.
Improved .to_json() and .from_json() to handle disp_cols and attrs. Fixed tests.
Added use of KeysDisabledError whenever a function attempts to lookup a row when keyfield is not set.
    In contrast, if keyfield is set and key not found, then KeyError if silent_error False.
    Fixed unit tests.